Задача D - Топологическая сортировка
