--- Traffic_light.cpp
@@ -1,103 +1,0 @@
-#include<iostream>
-#include<vector>
-
-
-// "Traffic light" task, http://acmp.ru/index.asp?main=task&id_task=124
-
-class Graph
-{
-public:
-	typedef size_t Vertex;
-	Graph(size_t vertex_count, bool is_directed)
-	{
-		vertex_count_ = vertex_count;
-		is_directed_ = is_directed;
-	}
-
-	size_t getVertexCount() const
-	{
-		return vertex_count_;
-	}
-
-	size_t getEdgeCount() const
-	{
-		return edge_count_;
-	}
-
-	virtual std::vector<Vertex> getNeighbors(const Vertex& v) const = 0;
-
-	virtual void addEdge(const Vertex& start, const Vertex& finish)
-	{
-		edge_count_++;
-	}
-	virtual size_t getNeighborsCount(const Vertex& v) const = 0;
-
-protected:
-	bool is_directed_;
-	size_t vertex_count_,
-		edge_count_;
-};
-
-
-class GraphAdjList : public Graph
-{
-
-public:
-	GraphAdjList(const size_t vertex_count, bool is_directed) :
-		Graph(vertex_count, is_directed)
-	{
-		adj_list_ = std::vector<std::vector<Vertex>>(vertex_count);
-	}
-
-	void addEdge(const Vertex& start, const Vertex& finish)
-	{
-		Graph::addEdge(start, finish);
-		adj_list_[start].push_back(finish);
-		if (!is_directed_)
-			adj_list_[finish].push_back(start);
-	}
-
-	std::vector<Vertex> getNeighbors(const Vertex& v) const override
-	{
-		return adj_list_[v];
-	}
-
-	size_t getNeighborsCount(const Vertex&  v) const override
-	{
-		return adj_list_[v].size();
-	}
-
-	void printGraph() const
-	{
-		std::cout << "Print\n";
-		std::cout << vertex_count_;
-		for (size_t i = 0; i < vertex_count_; i++)
-		{
-			for (auto k : adj_list_[i])
-				std::cout << k << " ";
-			std::cout << std::endl;
-		}
-	}
-
-private:
-	std::vector<std::vector<Vertex>> adj_list_;
-};
-
-int main()
-{
-	int n, m;
-	std::cin >> n;
-	std::cin >> m;
-	GraphAdjList g(n, false);
-	for (size_t i = 0; i < m; i++)
-	{
-		size_t a, b;
-		std::cin >> a;
-		std::cin >> b;
-		g.addEdge(a - 1, b - 1);
-	}
-	for (int i = 0; i < n; i++)
-	{
-		std::cout << g.getNeighborsCount(i) << " ";
-	}
-}
